refactor: test this rule header section
